+++ src/core/prelude/object/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.24.0 (2020-08-12)
+
+#### :rocket: New Feature
+
+* Added `Object.delete`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature